+++ src/HaruFont.php
@@ -9,7 +9,7 @@ class HaruFont
 
     public function __construct($font_ref)
     {
-        $this->ffi = \FFI::load(__DIR__.'/hpdf.h');
+        $this->ffi = HaruFFI::get_ffi();
         $this->h = $font_ref;
         if(is_null($this->h)) {
             throw new HaruException('Cannot create HaruFont handle');