+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Require ext-ffi.
+- Add HaruFFI abstraction so we only call FFI load once.
 
 ## [0.2.0] - 2023-04-15